+++ packages/generator/src/files/generate-template-baseLocale.ts
@@ -1,30 +1,36 @@
 import { join } from 'path'
 import { writeFileIfContainsChanges, writeFileIfNotExists } from '../file-utils'
 import { GeneratorConfigWithDefaultValues } from '../generate-files'
-import { sanitizeLocale } from '../generator-util'
+import { prettify, sanitizeLocale } from '../generator-util'
+import { fileEnding, importTypes, jsDocImports, jsDocType, shouldGenerateJsDoc, tsCheck, type } from '../output-handler'
 
-const getBaseLocaleTemplate = (baseLocale: string, importType: string) => {
+const getBaseLocaleTemplate = (baseLocale: string) => {
 	const sanitizedLocale = sanitizeLocale(baseLocale)
 
-	return `${importType} { BaseTranslation } from 'typesafe-i18n'
+	return `${tsCheck}
+${importTypes('typesafe-i18n', 'BaseTranslation')}
 
-const ${sanitizedLocale}: BaseTranslation = {
+${jsDocImports(
+		{ from: `typesafe-i18n`, type: 'BaseTranslation' },
+	)}
+
+${jsDocType('BaseTranslation')}
+${shouldGenerateJsDoc ? 'module.exports' : `const ${sanitizedLocale}${type('BaseTranslation')}`} = {
 	// TODO: your translations go here
 }
 
-export default ${sanitizedLocale}
+${shouldGenerateJsDoc ? '' : `export default ${sanitizedLocale}`}
 `
 }
 
 export const generateBaseLocaleTemplate = async (
 	config: GeneratorConfigWithDefaultValues,
-	importType: string,
 	forceOverride: boolean,
 ): Promise<void> => {
 	const { outputPath, baseLocale } = config
 
-	const baseLocaleTemplate = getBaseLocaleTemplate(baseLocale, importType)
+	const baseLocaleTemplate = getBaseLocaleTemplate(baseLocale)
 
 	const write = forceOverride ? writeFileIfContainsChanges : writeFileIfNotExists
-	await write(join(outputPath, baseLocale), `index.ts`, baseLocaleTemplate)
+	await write(join(outputPath, baseLocale), `index${fileEnding}`, prettify(baseLocaleTemplate))
 }